requirements.txt
streamlit
transformers
llama-index
elasticsearch
python-dotenv
openai